Added categoryFilter : useful for cookie categories
Load a tag (& script) only it is belongs to the categoryFilter. That means that for the Cookie Law, it's possible to easily classify tags by categories such as technical (1), analytics(2), advertisement(3) and social(4).. or even more sub categories. So if you decide to prevent a set of tag to load, its easier, just modify the categoryFilter array.